Modify C_05

+++ C_05/ex06/ft_is_prime.c
@@ -12,7 +12,7 @@
 
 int	ft_is_prime(int nb)
 {
-	int	i;
+	long long	i;
 
 	if (nb <= 1)
 		return (0);
